+++ __mocks__/@vtex/diagnostics-semconv.ts
@@ -4,6 +4,8 @@ const ATTR_VTEX_IO_WORKSPACE_NAME = 'vtex_io.workspace.name'
 const ATTR_VTEX_IO_WORKSPACE_TYPE = 'vtex_io.workspace.type'
 const ATTR_VTEX_IO_APP_ID = 'vtex_io.app.id'
 const ATTR_VTEX_IO_APP_AUTHOR_TYPE = 'vtex_io.app.author-type'
+const ATTR_VTEX_IO_CLUSTER_ID = 'vtex_io.cluster.id'
+const ATTR_VTEX_IO_CLUSTER_ROLE = 'vtex_io.cluster.role'
 
 export {
   ATTR_VTEX_ACCOUNT_NAME,
@@ -11,4 +13,6 @@ export {
   ATTR_VTEX_IO_WORKSPACE_TYPE,
   ATTR_VTEX_IO_APP_ID,
   ATTR_VTEX_IO_APP_AUTHOR_TYPE,
+  ATTR_VTEX_IO_CLUSTER_ID,
+  ATTR_VTEX_IO_CLUSTER_ROLE,
 }